worker/runtime: remove no-op device rule
We did a bunch of testing to verify that this rule is truly a no-op. We
even wrote a garbage rule:

```
{Allow: false, Access: "akdsjasdjasd"},
```

and the container was created successfully. With or without that rule
the list of devices in
/sys/fs/cgroup/devices/<container-id>/devices.list was the same as the
unprivileged containers under
/sys/fs/cgroup/devices/garden/<container-id>/devices.list

We later learned that this default list of devices is very permissive
and "arguably insecure" in the words of a runc dev.

In future it might make sense to further restrict unprivileged
containers, specifically removing the first two rules which allow mknod
for all block and char devices, which basically gives you read-write for
any device...

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>